Implemented Transactions in dbUpdateTable
The Update Table funtion was not using transactons when deleateing and then updateing records in the db.  This could have lead to the row being delated and then lost if the connection was lost durin the function.  To test is use the code below, killing the connection to the db between the deleat and the write.

```
### Use DBI transcation so that the row is deleated iff the row is then re-inserted
  DBI::dbWithTransaction(
    conn = con,
    code = {
      dbDeleteRowByKey(con, name, dt)
      if(verbose) message("Deleating row(s) form database")

      Sys.sleep(120) # Time to kill the connection

      RMySQL::dbWriteTable(con, name, dt, row.names = FALSE, append = TRUE)
      if(verbose) message("Writing row(s) to database")
    }
  )
```

Also changed the verbose messages to make them even more verboser.