MySQL text type added, and long_text renamed
There is a MySQL data type called text that allows for text of
inditermnat lenght to be saved in the db.  This meant that having
a data type called text_long that became a VARCHAR(30000) could
be confusing. So I have changed the name of long_text to
character_long.  I think that I was the only person to be using the
long_text funciton so i doubt that this will cause any problems.

The acctual addition in this is the ability to use the MySQL text
class in dbUpdate table.  There is now a funciton in dbUpdateTable,
text, that will create a sting of class text that is then used to
create a row of class text in the db.

The reason that I would like to add this class to dbUpdateTable is
so that strings of an arbitary length can be stored in the db.